Allow compiling against "dev" version of Coq
This matches e.g. Iris allowing a "dev" version of std++: You can install a
"dev" version to test stuff, but then you are responsible for making sure that
these versions actually work together.  We rely on that when testing things
against Iris master every night, for which purpose we install Iris master as
"dev" version.